prnt_list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t i = 0;
+	unsigned int i = 0;
 	list_t *ptr = NULL;
 
 	ptr = h;
